+++ src/main/java/com/dku/springstudy/dto/SignupDTO.java
@@ -1,11 +1,11 @@
-package com.dku.springstudy.model;
+package com.dku.springstudy.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class MemberForm {
+public class SignupDTO {
     private String email;
     private String password;
     private String name;
